refactor prompt shared to non server
