+++ packages/shared/src/index.ts
@@ -38,6 +38,9 @@ export * from "./features/comments/types";
 // experiments
 export * from "./features/experiments/utils";
 
+// prompts
+export * from "./features/prompts/parsePromptDependencyTags";
+
 // export db types only
 export * from "@prisma/client";
 export { type DB } from "../prisma/generated/types";
