+++ web/src/features/prompts/server/routers/promptRouter.ts
@@ -25,6 +25,7 @@ import {
   getObservationsWithPromptName,
   getObservationMetricsForPrompts,
   getAggregatedScoresForPrompts,
+  buildAndResolvePromptGraph,
 } from "@langfuse/shared/src/server";
 import { aggregateScores } from "@/src/features/scores/lib/aggregateScores";
 
@@ -775,6 +776,34 @@ export const promptRouter = createTRPCRouter({
           ),
           traceScores: aggregateScores(promptTraceScores?.scores ?? []),
         };
+      });
+    }),
+  resolvePromptGraph: protectedProjectProcedure
+    .input(
+      z.object({
+        promptId: z.string(),
+        projectId: z.string(),
+      }),
+    )
+    .query(async ({ input, ctx }) => {
+      const { promptId, projectId } = input;
+
+      throwIfNoProjectAccess({
+        session: ctx.session,
+        projectId,
+        scope: "prompts:read",
+      });
+
+      const prompt = await ctx.prisma.prompt.findUniqueOrThrow({
+        where: {
+          id: promptId,
+          projectId,
+        },
+      });
+
+      return buildAndResolvePromptGraph({
+        projectId: input.projectId,
+        parentPrompt: prompt,
       });
     }),
 });
